Solution with HashRouter
