+++ src/components/NotFoundPage/NotFoundPage.tsx
@@ -1,0 +1,3 @@
+export const NotFoundPage = () => (
+  <h1 className="todoapp__title">Page not found</h1>
+);
