+++ src/index.tsx
@@ -1,22 +1,16 @@
 import { createRoot } from 'react-dom/client';
+import { HashRouter } from 'react-router-dom';
 
 import 'bulma/css/bulma.css';
 import '@fortawesome/fontawesome-free/css/all.css';
 import './styles/index.scss';
 
 import { App } from './App';
-import { AuthProvider } from './components/Auth/AuthContext';
-import { DefaultProvider } from './components/DefaultContext';
-import { ErrorProvider } from './components/ErrorNotification/ErrorContext';
 
 const Root = () => (
-  <AuthProvider>
-    <ErrorProvider>
-      <DefaultProvider>
-        <App />
-      </DefaultProvider>
-    </ErrorProvider>
-  </AuthProvider>
+  <HashRouter>
+    <App />
+  </HashRouter>
 );
 
 createRoot(document.getElementById('root') as HTMLDivElement)
