--- src/types/Status.ts
@@ -1,5 +1,0 @@
-export enum Status {
-  ALL = 'All',
-  ACTIVE = 'Active',
-  COMPLETED = 'Completed',
-}
